Routing is done, static is done, good place to end
The settings.py was a bit tricky especially since they changed it from
1.7 to 1.8 so there were not that many resources onlince about it. css
is ready to be implemented and subsequently js. jquery will be good in
this bad boy. next up is database completion and login and regestration.
wonder if django has a similar email engine to flask......we'll find out